Add compilation pipeline orchestrator
